+++ test/dsl_test.rb
@@ -3,8 +3,8 @@
 class DslTest < Minitest::Test
   def setup
     @original_money_defined = Object.const_defined?(:Money)
+    @original_money = Object.const_get(:Money) if @original_money_defined
     @original_currency_defined = Object.const_defined?(:Currency)
-    @original_money    = Object.const_get(:Money)    if @original_money_defined
     @original_currency = Object.const_get(:Currency) if @original_currency_defined
   end
 
@@ -13,33 +13,32 @@ class DslTest < Minitest::Test
     restore_const(:Currency, @original_currency_defined, @original_currency)
   end
 
-  def test_defines_top_level_aliases_when_missing
-    remove_const(:Money)
+  def test_money_is_auto_bound_after_require
+    assert Object.const_defined?(:Money), 'Money should be defined by require minting'
+    assert_equal Mint::Money, Money
+  end
+
+  def test_currency_is_not_auto_bound
+    refute Object.const_defined?(:Currency),
+           'Currency should not be auto-bound; require minting/mint/aliases to opt in'
+  end
+
+  def test_aliases_binds_currency
     remove_const(:Currency)
 
-    Mint.use_top_level_constants!
+    require 'minting/mint/aliases'
 
-    assert_equal Mint::Money, Money
+    assert Object.const_defined?(:Currency)
     assert_equal Mint::Currency, Currency
   end
 
-  def test_does_not_override_existing_money
-    Object.const_set(:Money, :existing_money)
+  def test_aliases_warns_when_currency_already_defined
+    remove_const(:Currency)
     Object.const_set(:Currency, :existing_currency)
 
-    assert_raises(NameError) { Mint.use_top_level_constants! }
-    assert_equal :existing_money, Money
+    assert_output('', /already defined/) { load_aliases_file }
+
     assert_equal :existing_currency, Currency
-  end
-
-  def test_does_not_override_already_defined_money
-    Object.const_set(:Money, Mint::Money)
-    Object.const_set(:Currency, Mint::Currency)
-
-    assert_output('', /already defined/) { Mint.use_top_level_constants! }
-
-    assert_equal Mint::Money, Money
-    assert_equal Mint::Currency, Currency
   end
 
   private
@@ -51,5 +50,11 @@ class DslTest < Minitest::Test
   def restore_const(name, was_defined, value)
     remove_const(name)
     Object.const_set(name, value) if was_defined
+  end
+
+  # Re-loads aliases.rb to exercise the already-defined branch without
+  # re-triggering the require guard (require returns false on second load).
+  def load_aliases_file
+    load File.expand_path('../lib/minting/mint/aliases.rb', __dir__)
   end
 end